requirements.txt
beautifulsoup4==4.4.0
Django==1.8.4
django-appconf==1.0.1
django-compressor==1.5
django-model-utils==2.3.1
django-modelcluster==0.6.2
django-taggit==0.16.3
django-treebeard==3.0
ecdsa==0.13
Fabric==1.10.2
gunicorn==19.3.0
html5lib==0.999
paramiko==1.15.2
Pillow==2.9.0
pycrypto==2.6.1
pytz==2015.4
requests==2.7.0
six==1.9.0
sorl-thumbnail==12.3
Unidecode==0.4.18
wagtail==1.0
wheel==0.24.0
Willow==0.2.1
